Add cropping
Add cropping to xrm2norm_manyimgforangle.

This script allows to go from individual xrm images to a single
tomography.

In this case, the xrm contains many images for each angle, which
are normalized, cropped, and then an average is done for all
images with the same angle.

At the end a single normalized cropped tomography remains.