Complete script for alignment and fusion
Script used to align and fuse multiple images for the same
angle in mrc tomography projection files.

All the projections for the same angle are aligned using TomoJ
FTEM, then averaged (by adding and dividing by the number of
images). This is done for all the images of the same angle of
different tomography raw data.