Begin transition to use new xrm2nexus
Rename to old the scripts using manytomos2nexus (which is
the old xrm2nexus).

manytomos2nexus uses the date of creation of the file, and uses
too much the name of the file for preprocessing. This is not robust,
and prone to errors: file naming, copy and paste with the subsequent
lost of date of creation of files which makes the preprocessing
unusable...

We keep the old scripts with the suffix _old.