Continue transition to use new xrm2nexus
The new xrm2nexus is using a tinyDB database to index files
based on the TXM txt file. This allows to avoid some
problems with parsing of dataname for finding Energy, ZP...
It also allows to get rid of problems with data of creation of
files, that was used before to organize files inside hdf5,
and could create misordered hdf5 with angles for example, after
a copy and paste of files that was changing the date of creation
of files. Then the copy with permits (cp -p ...) was totally
required to do not lose the date of creation of acquired xrm
images. With the new scripts, which use xrm2nexus instead of
manytomos2nexus after this transition,
this should not be anymore the case, and they should work
independently of the date and time of creation of the acquired
images.

This script is used for magnetism thin layers experiments
(sferrer experiments), using many repetitions for each of the
images.