Add tomo_deconv for exporting variable
LD_LIBRARY_PATH has to be exported in order to get the Matlab
runtime in pcbl0906.

This export is better to be here in the script, rather than
to be exported at the system or user level, where it caused
problems with other SW, as it is the case of the library
cv2.

Thus we have created this script which calls inside it
the tomo_deconv_wiener.