New version: from bash to python
Add new version of fusion mrc. Migrated to python.

New feature allowing to merge only the files related to the
same sample.

Different samples can be located in the same directory, and
the fusion will be done by files related to same samples.

Parsing is done by filename. It is possible to indicate how many
fields of the filename have to match to be considered in the
same group of files for merging. Default is 2 (normally: date +
sample_name). 3 could be used for example if the groups have to
be done by: date + sample_name + energy.